+++ back/src/domains/convention/use-cases/notifications/NotifyLastSigneeThatConventionHasBeenSigned.unit.test.ts
@@ -1,4 +1,5 @@
 import {
+  AgencyWithUsersRights,
   ConventionDto,
   ConventionDtoBuilder,
   errors,
@@ -10,7 +11,6 @@ import {
   ExpectSavedNotificationsAndEvents,
   makeExpectSavedNotificationsAndEvents,
 } from "../../../../utils/makeExpectSavedNotificationAndEvent.helpers";
-import { AgencyWithUsersRights } from "../../../agency/ports/AgencyRepository";
 import { makeSaveNotificationAndRelatedEvent } from "../../../core/notifications/helpers/Notification";
 import { CustomTimeGateway } from "../../../core/time-gateway/adapters/CustomTimeGateway";
 import { InMemoryUowPerformer } from "../../../core/unit-of-work/adapters/InMemoryUowPerformer";
